IoT-Platform/Layer2Data/dashboard/requirements.txt
backports.ssl-match-hostname==3.5.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
docker==2.7.0
docker-pycreds==0.2.1
Flask==0.12.2
future==0.16.0
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.3.5
MarkupSafe==1.0
python-telegram-bot==9.0.0
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
Werkzeug==0.14.1
